% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_3rm_1f.R
\name{agreguj_cawi_r3_f1}
\alias{agreguj_cawi_r3_f1}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
agreguj_cawi_r3_f1(wskazniki, grupy)
}
\arguments{
\item{wskazniki}{ramka danych z wynikami 3. rundy monitoringu}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie zagregowanym
na podstawie ramki danych z wynikami ankiety CAWI.
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{dane_szkoly}},}
\item{\code{\link{firma_bad}},}
\item{\code{\link{l_abs}},}
\item{\code{\link{l_abs_WT}},}
\item{\code{\link{l_kobiet}},}
\item{\code{\link{l_kobiet_WT}},}
\item{\code{\link{formy}},}
\item{\code{\link{zawod_licz}},}
\item{\code{\link{zawod_licz_WT}},}
\item{\code{\link{W1_pon_wyb}},}
\item{\code{\link{W1_pon_wyb_WT}},}
\item{\code{\link{W2_przyg_zaw}},}
\item{\code{\link{W2_przyg_zaw_WT}},}
\item{\code{\link{W2_zawod}},}
\item{\code{\link{W2_zawod_WT}},}
\item{\code{\link{NZ2_pnz_szk}},}
\item{\code{\link{NZ2_pnz_szk_WT}},}
\item{\code{\link{NZ2_pnz_prac}},}
\item{\code{\link{NZ2_pnz_prac_WT}},}
\item{\code{\link{PL1_plany_eduk}},}
\item{\code{\link{PL1_plany_eduk_WT}},}
\item{\code{\link{PL2_plany_praca}},}
\item{\code{\link{PL2_plany_praca_WT}},}
\item{\code{\link{PL5A_praca_branza}},}
\item{\code{\link{PL5A_praca_branza_WT}},}
\item{\code{\link{PL5B_praca_branza_plan}},}
\item{\code{\link{PL5B_praca_branza_plan_WT}},}
\item{\code{\link{PL1_poza}},}
\item{\code{\link{PL1_poza_WT}}}
}
}
