% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{firma_bad}
\alias{firma_bad}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
firma_bad(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca informację o nazwie firmy realizującej
badanie.
}
