% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{l_abs}
\alias{l_abs}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
l_abs(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zbadanych absolwentów
w grupie.
}
