% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{l_kobiet}
\alias{l_kobiet}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
l_kobiet(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
liczba
#' @importFrom dplyr \%>\% filter count pull
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zbadanych kobiet w
grupie.
}
