% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{l_kobiet_WT}
\alias{l_kobiet_WT}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
l_kobiet_WT(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
liczba
}
\description{
Funkcja oblicza \strong{ważony} wskaźnik opisujący liczbę
zbadanych kobiet w grupie.
}
