% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{NZ2_pnz_prac}
\alias{NZ2_pnz_prac}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
NZ2_pnz_prac(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{ocena praktycznej nauki zawodu
w szkole}: "Jak Pani / Pan ocenia, w jakim stopniu praktyczna nauka zawodu
przygotowała Panią/Pana do wykonywania pracy w zawodzie? - u pracodawcy w
Polsce".
Wskaźnik jest rozkładem odpowiedzi:
\itemize{
\item{\code{t2b}: "Przygotowała dobrze" + "Przygotowała bardzo dobrze" - kod
4 + kod 5}
\item{\code{wcale}: "Nie przygotowała w ogóle" - kod 1}
\item{\code{slabo}: "Przygotowała słabo" - kod 2}
\item{\code{srednio}: "Przygotowała średnio" - kod 3}
\item{\code{dobrze}: "Przygotowała dobrze" - kod 4}
\item{\code{b_dob}: "Przygotowała bardzo dobrze" - kod 5}
\item{\code{trud_pow}: "Trudno powiedzieć" - kod 7}
}
}
