% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{PL1_plany_eduk}
\alias{PL1_plany_eduk}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
PL1_plany_eduk(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{planowanie dalszej edukacji}
będący odsetkiem uczniów, którzy w pytaniu "PL1.: Proszę pomyśleć o swoich
planach na pierwszy rok po ukończeniu obecnej szkoły. Czy po zakończeniu
nauki w obecnej szkole planuje Pani/Pan:...", w podpunktach:
\itemize{
\item{... uczyć się w szkole branżowej drugiego stopnia}
\item{... uczyć się w liceum dla dorosłych}
\item{... uczyć się w szkole policealnej}
\item{... studiować}
\item{... uczyć się na kwalifikacyjnym kursie zawodowym (KKZ), czyli kursie,
który kończy się egzaminem zawodowym}
}
zaznaczyli chociaż jedną odpowiedź "Tak, taki mam plan" (kod 1.)
}
