% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{PL1_poza}
\alias{PL1_poza}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
PL1_poza(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{pozostawanie poza edukacją i
rynkiem pracy}, czyli odsetki i liczebności uczniów, którzy pozostają poza
edukacją i rynkiem pracy - na podstawie pytania PL1
(\code{\link{PL1_plany_eduk_WT}})
}
