% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{PL2_plany_praca_WT}
\alias{PL2_plany_praca_WT}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
PL2_plany_praca_WT(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca \strong{ważony} wskaźnik \strong{planowanie
poszukiwania pracy} będący odsetkiem uczniów, którzy w pytaniu "PL2.: Czy w
pierwszym roku po zakończeniu nauki w obecnej szkole planuje Pani/Pan
poszukiwać pracy?".
}
