% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{PL5A_praca_branza}
\alias{PL5A_praca_branza}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
PL5A_praca_branza(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{planowanie zatrudnienia w
branży}, czyli odsetek uczniów, którzy podejmą pracę w branży związanej
z wyuczonym zawodem w ciągu roku od ukończenia nauki - pytanie PL5A kod 1.
}
