% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{PL5B_praca_branza_plan}
\alias{PL5B_praca_branza_plan}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
PL5B_praca_branza_plan(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{planowanie zatrudnienia w
branży}, czyli odsetek uczniów, którzy planują podjąć pracę w branży
związanej z wyuczonym zawodem w ciągu roku od ukończenia nauki - pytanie
PL5B_1 kod 1.
}
