% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{W2_przyg_zaw}
\alias{W2_przyg_zaw}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
W2_przyg_zaw(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja licząca wskaźnik \strong{ocena szkoły - porzygotowanie
do wykonywania zawodu}: "Jak  Pani/Pan uważa, na ile szkoła przygotowała
Panią/Pana do wykonywania zawodu, którego się Pani/Pan w niej uczy?".
Wskaźnik jest rozkładem odpowiedzi:
\itemize{
\item{\code{t2b}: "Przygotowała dobrze" + "Przygotowała bardzo dobrze" - kod
4 + kod 5}
\item{\code{b2b}: "Nie przygotowała w ogóle" + "Przygotowała słabo" - kod 1
\itemize{
\item kod 2}
\item{\code{wcale}: "Nie przygotowała w ogóle" - kod 1}
\item{\code{slabo}: "Przygotowała słabo" - kod 2}
\item{\code{srednio}: "Przygotowała średnio" - kod 3}
\item{\code{dobrze}: "Przygotowała dobrze" - kod 4}
\item{\code{b_dob}: "Przygotowała bardzo dobrze" - kod 5}
\item{\code{trud_pow}: "Trudno powiedzieć" - kod 7}
}
}
}
