% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{W2_zawod}
\alias{W2_zawod}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
W2_zawod(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek uczniów oceniających, że szkoła
przygotowanła ich dobrze lub doskonale do wykonywania zawodu (\emph{W1}).
}
