% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_3rm_1f.R
\name{zawod_licz}
\alias{zawod_licz}
\title{Obliczanie wskaznikow dla R3 F1 na poziomie zagregowanym}
\usage{
zawod_licz(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca liczbę absolwentów danego zawodu. Ponadto,
dodana jest informacja o branży, do której należy zawód.
}
